refactor(pots): add serialization support and transient properties to PokerPots and Pot classes
