+++ src/main/kotlin/hwr/oop/projects/peakpoker/core/pot/PokerPots.kt
@@ -2,11 +2,13 @@ package hwr.oop.projects.peakpoker.core.pot
 
 import hwr.oop.projects.peakpoker.core.card.CommunityCards
 import hwr.oop.projects.peakpoker.core.player.PokerPlayer
+import kotlinx.serialization.Serializable
 
+@Serializable
 class PokerPots(
   private val players: List<PokerPlayer>,
   private val communityCards: CommunityCards,
-  private val pots: MutableList<Pot> =
+  val pots: MutableList<Pot> =
     mutableListOf(
       Pot(
         eligiblePlayers = players.toSet(),
@@ -15,7 +17,7 @@ class PokerPots(
     ),
 ) : Iterable<Pot> by pots {
   private val mainPot: Pot get() = pots.first()
-  private val totalPotAmount: Int get() = pots.sumOf { it.amount() } // TODO: Do we need this?
+  // private val totalPotAmount: Int get() = pots.sumOf { it.amount() } TODO: Do we need this?
 
   fun addChipsToMainPot(chips: Int) {
     mainPot.addChips(chips)
